Adds placeholders for 'src/ast' and 'tokens::collect_invalid'
Unfortunately it appears that the parsing exmaple may be broken on a few
fronts. It produces an unexpected error (an InvalidChar token on what
should be valid input) and fails to provide the correct byte index in
the source. The first of these will be fixed in subsequent commits. The
second will not.