Fixes both previously described issues; change token indices
The issues were described in the previous commit. They have both been
fixed.

Notable changes:
Produced tokens now properly give their starting byte indices in the file, instead of their indices relative to their containing token (if they are part of a block). This change was made to allow the printing of individual invalid tokens once they are collected, without the context of the tokens that contain them.