Add different types of values to `exec`; change ints to `isize`
This commit replaces the `exec::Value` type alias with an enum of the
same name containing three variants: `Unit`, `Bool`, and `Int`. The
first is the "unit type", or the type that is given as the value of an
empty expression or function that does not return a value. The meaning
of values of type `Bool` or `Int` should be fairly self-explanatory.

Additionally, the integer type previously used was a `usize`. This has
now been changed (per previous discussion) to an `isize`.

A bug was discovered during the process of making this change - it has
been filed as issue #9.