Improved optimiser performance by ~6x
This is achieved by keeping the information regarding which bounds
origionates from the same requirements. Then, when we make a
substitution, we don't give the child any of the bounds from the same
requirement since substituting a bound from the same requirement is
effectively just a change in variables that would have been acheived by
using that substitution in the first place.
This decreases the width of the optimisers trees significantly.