Fix unsoundness issue in `Inequality` creation
An edge case that was previously left unaddressed was to do with how
`Inequality`s are created from `Requirement`s: Because the
representation in `Inequality` has no denominators, all terms are
multiplied out, where an aggregate expression is used as the value that
both sides of the inequality are multiplied by.

If this expression is negative, however, the direction of the inequality
should actually be reversed! This was not tracked previously, but this
indicates that we need to know the sign of the expression before
providing an `Inequality` into the prover.

The details of how this is resolved are provided in the changes from
this commit, so it will not be dicussed further here, other than stating
that this has now been addressed.