Verification for proof statments
This is the initial work, and most of it is thrown together in a rather
haphazard manner, but it appears to work in most cases. The bulk of the
modification is within 'src/verify/mod.rs' and a new file,
'src/verify/prover.rs'.

An additional test file - 'test_input_fails_graph.tc' - has been added,
using an example that fails to compile with the graph method but will
compile successfully with the bounds method. This will likely only
require a small change to fix, but it will be done separately.

This commit also introduces abundant usage of `unsafe` in order to get
around the borrow-checker. There are a couple patterns worth describing
here. The first is used in two places: firstly for storing provers
alongside errors, because errors may sometimes borrow content from the
prover, and secondly for storing the dependent provers inside
`verify::prover::WrappedProver`. More details on the latter are in the
type's documentation. The former is largely irrelevant outside calls to
`TopLevelScope::{check_fn,check_items}`.

The second pattern worth mentioning is how the central prover is passed
around during verification within a single function. It is represented
as an `Option<&mut WrappedProver>`, but only using this limits the
functionality because it cannot be cloned - only moved into wherever
it's used. To get around this, the common pattern is to turn this type
into an `Option<*mut WrappedProver>` at the top of the function, and
then generate the original type whenever it is needed, taking care not
to allow any aliased mutable references (though this is usually not an
issue, as - as mentioned - usage of the type will move it).

As a final note, the test input file currently *does not* compile, due
to a single proof failure, and I have not investigated precisely why
this is occuring (whether it's operator error or a bug).