Misc cleanup, removing warnings
There were *many* warnings about unused imports, mutability on
variables, and such. This gets rid of some of those. The implementation
of the bounds prover was left alone because there seem to be many pieces
that are currently a work in progress.

The only notable change is the removal of 'proof/error.rs', which was
unused. It was previously made as part of planning to have proof
checking as separate from 'src/verify', but that's since changed.