Clean up prover handling in 'verify'
The primary change was to switch out all occurances of
`Option<&mut Provers>` for `&mut Provers`, adding the `mask` field to
`Provers` to account for cancelling proof instead. The motivation here
was to properly integrate masking in as a replacement for the single
on/off dynamic given by using `Option`.